Added functionality to display a new hand. Or hand from the database

+++ src/main/java/com/blackjack/repository/HandRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.blackjack.entity.Hand;
 
-public interface HandRepository extends JpaRepository<Hand, Long>{
+public interface HandRepository extends JpaRepository<Hand, Integer>{
 
 }
